Refactor base client and transaction
