+++ pysui/sui/sui_pgql/execute/caching_txn.py
@@ -11,7 +11,7 @@ from typing import Any, Coroutine, Optional, Union
 from deprecated.sphinx import versionchanged
 from pysui.sui.sui_pgql.pgql_clients import AsyncSuiGQLClient
 from pysui.sui.sui_pgql.pgql_txb_signing import SignerBlock
-from pysui.sui.sui_pgql.pgql_txn_base import _TransactionBase as txbase
+from pysui.sui.sui_common.trxn_base import _TransactionBase as txbase
 
 from pysui.sui.sui_bcs import bcs
 import pysui.sui.sui_pgql.pgql_validators as tv
