+++ pysui/sui/sui_pgql/pgql_async_txn.py
@@ -10,7 +10,7 @@ import asyncio
 from typing import Any, Callable, Optional, Union
 from deprecated.sphinx import versionchanged, versionadded, deprecated
 from pysui.sui.sui_pgql.pgql_clients import SuiGQLClient
-from pysui.sui.sui_pgql.pgql_txn_base import _SuiTransactionBase as txbase
+from pysui.sui.sui_common.trxn_base import _SuiTransactionBase as txbase
 
 from pysui.sui.sui_bcs import bcs
 from pysui.sui.sui_txn.transaction_builder import PureInput
